# docs/swagger.yaml
basePath: /
definitions:
  github.com_miprokop_fication_internal_http-server_handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
  github.com_miprokop_fication_internal_http-server_handlers.eventAllData:
    properties:
      created_by_id:
        type: string
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_status:
        type: string
      event_type:
        type: string
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      organization_id:
        type: string
      staff:
        items:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.staffEvents'
        type: array
      steps:
        items:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.stepShortResponse'
        type: array
    type: object
...
  github.com_miprokop_fication_internal_http-server_handlers.eventRequestUpdate:
    properties:
      created_by_id:
        type: string
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_status:
        type: string
      event_type:
        type: string
      id:
        type: string
      image_path:
        type: string
      name:
        type: string
      organization_id:
        type: string
    type: object
...
  github.com_miprokop_fication_internal_http-server_handlers.organizationType:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  github.com_miprokop_fication_internal_http-server_handlers.staffEvents:
    properties:
      event_id:
        type: string
      id:
        type: string
      staff_id:
        type: string
      staff_role:
        type: string
      status:
        type: string
    type: object
...
  github.com_miprokop_fication_internal_http-server_handlers.stepShortResponse:
    properties:
      creation_date:
        type: string
      description:
        type: string
      end_date:
        type: string
      event_id:
        type: string
      id:
        type: string
      level:
        type: integer
      max_score:
        type: integer
      name:
        type: string
      status:
        type: string
      task:
        type: string
    type: object
...
  internal_http-server_handlers.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  internal_http-server_handlers.organizationResponse:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      positions:
        items:
          $ref: '#/definitions/internal_http-server_handlers.position'
        type: array
      teams:
        items:
          $ref: '#/definitions/internal_http-server_handlers.team'
        type: array
      types:
        items:
          $ref: '#/definitions/internal_http-server_handlers.organizationType'
        type: array
      website_url:
        type: string
    type: object
  internal_http-server_handlers.organizationType:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  internal_http-server_handlers.organizationUpdateResponse:
    properties:
      id:
        type: string
      image:
        type: string
      name:
        type: string
      website_url:
        type: string
    type: object
...
  internal_http-server_handlers.position:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  internal_http-server_handlers.team:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      organization_id:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: API Server for challenge service
  title: Acheer API
  version: "1.0"
paths:
...
  /api/event/:id:
    delete:
      consumes:
      - application/json
      description: |-
        delete event by id
        also delete chained steps
      operationId: delete-event
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Event By ID
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Get all invitations by current user
      operationId: get-invite-by-id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventAllData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get event by id
      tags:
      - events
    put:
      consumes:
      - application/json
      description: Update event by ID
      operationId: update-event
      parameters:
      - description: event
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventRequestUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.eventAllData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update event by id
      tags:
      - events
...
  /api/org/:
    get:
      consumes:
      - application/json
      description: Get All Organizations In Service
      operationId: get-all-organizations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_http-server_handlers.organizationResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            type: ""
        "403":
          description: Forbidden
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Organizations In Service
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: |-
        create models.Organization
        create organization
      operationId: create-organization
      parameters:
      - description: organization info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_http-server_handlers.organizationResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create organization
      tags:
      - organizations
  /api/org/:id:
    delete:
      consumes:
      - application/json
      description: |-
        delete organization by id
        also delete all chained staff, positions, steps, teams
      operationId: delete-organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Organization By ID
      tags:
      - organizations
    get:
      consumes:
      - application/json
      description: Get Organization By ID
      operationId: get-organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_http-server_handlers.organizationResponse'
        "400":
          description: Bad Request
          schema:
            type: ""
        "403":
          description: Forbidden
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Organization By ID
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update organization by id
      operationId: update-event
      parameters:
      - description: organization
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_http-server_handlers.organizationUpdateResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update organization by id
      tags:
      - organizations
...
  /api/org/type/:
    get:
      consumes:
      - application/json
      description: Get All Organizations Types In Service
      operationId: get-all-organizations-types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.organizationType'
            type: array
        "400":
          description: Bad Request
          schema:
            type: ""
        "403":
          description: Forbidden
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get All Organizations Types In Service
      tags:
      - organizations-type
    post:
      consumes:
      - application/json
      description: create models.OrganizationType
      operationId: create-organization-type
      parameters:
      - description: organization type info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.organizationType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create organization type
      tags:
      - organizations-type
  /api/org/type/:id:
    delete:
      consumes:
      - application/json
      description: delete organization type by id
      operationId: delete-organization
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Organization Type By ID
      tags:
      - organization-type
    get:
      consumes:
      - application/json
      description: Get Organization By ID
      operationId: get-organization-type
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.organizationType'
        "400":
          description: Bad Request
          schema:
            type: ""
        "403":
          description: Forbidden
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Organization Type By ID
      tags:
      - organizations-type
    put:
      consumes:
      - application/json
      description: Update organization type by id
      operationId: update-event
      parameters:
      - description: organization type
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.organizationType'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/github.com_miprokop_fication_internal_http-server_handlers.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update organization type by id
      tags:
      - organizations
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"